Fix platform integration tests
